+++ CHANGELOG.md
@@ -4,4 +4,6 @@ This log follows the conventions of
 
 ## Unreleased
 ### Added
+- Standard-library argparse integration.
+- New types for real numbers.
 - Documentation. This change log and an introduction by example.